Layout1/requirements.txt
boto3==1.14.30
botocore==1.17.30
click==7.1.2
dnspython==2.0.0
docutils==0.15.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
numpy==1.19.1
Pillow==7.2.0
pymongo==3.10.1
python-dateutil==2.8.1
s3transfer==0.3.3
scikit-image==0.17.2
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
